Initial changes with new idea
