+++ .env.dev
@@ -1,2 +1,2 @@
 FLASK_APP=project/__init__.py
-FLASK_ENV=development
+FLASK_DEBUG=development
