+++ Dockerfile
@@ -6,8 +6,11 @@ RUN apt install vim -y
 RUN python -m pip install --upgrade pip
 
 COPY requirements/ /tmp/requirements/
-COPY src/ /usr/app/src
+WORKDIR /usr/app
 
 RUN python -m pip install -r /tmp/requirements/base.txt
-WORKDIR /usr/app
+ENV FLASK_APP microblog.py
+
 EXPOSE 5000
+COPY microblog/ /usr/app
+CMD ["python", "-m", "app"]
