+++ docker-compose.yml
@@ -2,13 +2,8 @@ version: "3"
 services:
   web:
     build: .
-    command: python3 -m src
     ports:
       - 5000:5000
-    env_file:
-      - ./.env.dev
-    volumes:
-      - ./db/:/usr/app/db/
   # db:
   #   image: "postgres:13-alpine"
   #   environment:
@@ -17,5 +12,3 @@ services:
   #     POSTGRES_DB: postgres # The PostgreSQL default database (automatically created at first launch)
   #   volumes:
   #     - postgres_data/:/var/lib/postgresql/data/
-volumes:
-  db_data:
