+++ microblog/app/app.py
@@ -1,0 +1,6 @@
+from flask import Flask
+from .routes import router
+
+app = Flask(__name__)
+
+app.register_blueprint(router)
