+++ microblog/app/templates/base.html
@@ -15,15 +15,11 @@
             <nav class="navbar">
                 <div class="container">
                     <div id="navbarMenuHeroA" class="navbar-menu">
-                        <div class="navbar-end">
-                            {{ athlete_id }}
-                            <a href="{{ url_for('strava_authorize') }}" class="navbar-item">
-                                Strava Login
-                            </a>
-                        </div>
+                        </a>
                     </div>
                 </div>
-            </nav>
+        </div>
+        </nav>
         </div>
         <div class="hero-body">
             <div class="container has-text-centered">
