+++ microblog/app/templates/index.html
@@ -12,11 +12,8 @@
 </h3>
 
 <body>
-    <section class="hero is-fullheight" style="background-color:#f7f7f7;">
+    <section class="hero is-fullheight" style="background-color:#331616;">
         <div class="hero-body">
-            <a href="{{ url_for('strava_retreive_athlete') }}" class="navbar-item">
-                Strava Access Athlete Details
-            </a>
         </div>
     </section>
 </body>
