--- src/constants.py
@@ -1,2 +1,0 @@
-CLIENT_ID = ""
-CLIENT_SECRET = ""
